fix(ci): restore source line budget gates
