+++ src/tests/mod.rs
@@ -8,3 +8,4 @@ include!("chunk_07.rs");
 include!("chunk_08.rs");
 include!("chunk_09.rs");
 include!("chunk_10.rs");
+include!("chunk_11.rs");
